Switch columns editor to read-only during concurrent edits.
A concurrent edit can take place if a new widget is added to a column or if a pre-existing widget is deleted. In such a case altering the number of columns might lead to data loose. In this case it's better to stop allowing the editor to adjust columns then to loose data.

Fixes #414